[Kernel] define kernel download target/url with defined tags
In order to make sure the downloaded package is identical with package
that would be built by the source code, defining the download target
and url with defined tags.

So when a particular download target is not available yet, the download
will fail and user will have to build it.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>