eBPF helper function for attribute search in the netlink message
There are few network applications relying on Netlink subsystem to get
notifications for net-device attribute changes like MTU, Speed,
Oper-Status, Name, slave, slave info, etc. The Netlink subsystem
notifies the application on every attribute change regardless of what
is needed for the application. The attribute search support in
EBPF filter helps to filter the Netlink packets based on the specific
set of attributes that are needed for the application.

The classical BPF supports attribute search but that doesn't support
MAPS. The extended BPF supports MAPS, but the attribute search is not
enabled. Hence this patch enables the support for attribute search in
EBPF.

This patch adds the support for following helper function.
    FN(skb_get_nlattr),
    FN(skb_get_nlattr_nest)

skb_get_nlattr:
    Find a specific attribute in a stream of attributes

skb_get_nlattr_nest:
    Find a specific attribute in a stream of nested attributes