Enable MDIO related kernel config options
By default, MDIO_BUS_MUX is set indirectly via other configurations (in
particular OF must be enabled).  However, we don't need OF, but still
wanted MDIO_BUS_MUX to be configured for accessing required MDIO drivers.

The patch here is to remove the dependence on OF_MDIO, and add a prompt/label
to MDIO_BUS_MUX so that it can be explicitly configured.

Signed-off-by: Madhava Reddy Siddareddygari <msiddare@cisco.com>