Allow NULL of_node device
Cisco is using ACPI to read device configuration instead of
OF (device-tree) for ds4424 device.

Kernel is in a transition period in its support for ACPI
(cisco using ACPI for kernel module configuration).  The kernel
makes a distinction between device tree (of_node) and ACPI
(fw_node) in the device structure.  It is unclear if they can be
used together or not, but I can definitively say that we are only
using ACPI and not device tree.  In this particular driver, there
was an explicit check for having a device tree node, even though no
reference to the device tree was made within the driver.  This caused
the probe routine to fail with -ENODEV in cisco environment.
This patch removes the check for of_node.

Signed-off-by: Madhava Reddy Siddareddygari <msiddare@cisco.com>