[Mellanox] Remove incorrectly generated patch allegedly avoiding kernel deadlock but causes one
Patch 0028-thermal-Fix-deadlock-in-thermal-thermal_zone_device_.patch was incorrectly generated. It is supposed to backport Linux commit torvalds/linux@163b00c, but the commit was already backported to the stable series in v4.19.89, so included in Linux v.19.152 currently used by SONiC. Instead, the incorrect patch reverts the first hunk of the commit. So, remove it.

Regression tests have been performed against these patches on the Mellanox platform, no issue found.