NPU disabled used PCIe BAR
For Cisco ASIC only BAR0 is valid. Not disabling other BAR's
was resulting in pci_enable_device function failure in P0
Pacific ASIC's. Further debugging and consultion with Hardware
team, issue seems to be related to only P0 version of ASIC and
workaround suggested is to disable unused PCI BAR.

This is same pull request reviewed ealier,
https://github.com/Azure/sonic-linux-kernel/pull/225/

Cisco experimented removing this patch, but identified that
there are boards in the field with Pacific ASIC.

This patch is need for these ASIC's to work, hence re-submitting.

Signed-off-by: Madhava Reddy Siddareddygari <msiddare@cisco.com>