Re-add certain parts of cisco-x86-gpio-config.patch
In the 202012 branch, there is a config option committed to set the
number of GPIOs on the system to 10240, because some Cisco hardware
could have that many GPIOs. This config option is, for whatever reason,
not on the master branch.

Therefore, re-add cisco-x86-gpio-config.patch, but reduce it to the
parts that allow configuration of the config.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>