Update to 6.12.41 for Debian Trixie (#488)
Update the kernel to 6.12.41, part of the 6.12 series kernel used in Debian Trixie.

Additionally, with this, integrate SONiC-specific configs and patches into the build infra that Debian has set up for doing builds of different flavors and architectures of the kernel. There is room in that infra for downstream distributions to add in their own configs and patches. SONiC will use that portion of the infra to set custom kernel configs and include our patches there.

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Co-authored-by: Yan Markman <32263981+yanmarkman@users.noreply.github.com>
Co-authored-by: Tharigonda Rohini Kumar <rohini_kumar_tharigo@dell.com>
Co-authored-by: Hemanth Kumar Tirupati <tirupatihemanthkumar@gmail.com>